+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Scripting:
     * Python can now convert objects to path
     * More intuitive construction of C++ classes from python
+* Bug Fixes:
+    * Dragging keyframes properly preserves keyframe transitions
 
 ## 0.4.3